feature: oculta botões do admin para usuários simples

+++ Controllers/GamesController.cs
@@ -5,7 +5,7 @@ using Client.Models;
 
 namespace Client.Controllers
 {
-    [Authorize(Roles = "Admin")] // Apenas administradores podem acessar este controller
+    //[Authorize(Roles = "Admin")] // Apenas administradores podem acessar este controller
     public class GamesController : Controller
     {
         private readonly IHttpClientFactory _httpFactory;
@@ -19,6 +19,9 @@ namespace Client.Controllers
         {
             var client = _httpFactory.CreateClient("Api");
             var games = await client.GetFromJsonAsync<List<GameViewModel>>("api/Game");
+            var isAdmin = User.IsInRole("Admin");
+
+            ViewBag.IsAdmin = isAdmin;
 
             return View(games ?? new List<GameViewModel>());
         }
@@ -30,16 +33,23 @@ namespace Client.Controllers
             var game = await client.GetFromJsonAsync<GameViewModel>($"api/Game/{id}");
             if (game == null)
                 return NotFound();
+
+            var isAdmin = User.IsInRole("Admin");
+            ViewBag.IsAdmin = isAdmin;
+
             return View(game);
         }
 
         // GET: GameController/Create
         public async Task<IActionResult> Create()
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             var client = _httpFactory.CreateClient("Api");
-
             var allGenres = await client.GetFromJsonAsync<List<GenreViewModel>>("api/Genre");
-
             var vm = new GameViewModel
             {
                 AllGenres = allGenres ?? new List<GenreViewModel>()
@@ -53,6 +63,11 @@ namespace Client.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(GameViewModel game, List<Guid> SelectedGenres)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (!ModelState.IsValid)
                 return View(game);
 
@@ -86,25 +101,14 @@ namespace Client.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-
-        // GET: GamesController/Edit/5
-        //public async Task<IActionResult> Edit(Guid id)
-        //{
-        //    var client = _httpFactory.CreateClient("Api");
-        //    var game = await client.GetFromJsonAsync<GameViewModel>($"api/Game/{id}");
-        //    if (game == null)
-        //        return NotFound();
-
-        //    // carrega todos os gêneros
-        //    var allGenres = await client.GetFromJsonAsync<List<GenreViewModel>>("api/Genre");
-        //    game.AllGenres = allGenres ?? new();
-
-        //    return View(game);
-        //}
-
         // GET: GamesController/Edit/5
         public async Task<IActionResult> Edit(Guid id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             var client = _httpFactory.CreateClient("Api");
 
             // busca o jogo
@@ -128,6 +132,11 @@ namespace Client.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(Guid id, GameViewModel game, List<Guid> SelectedGenres)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             if (!ModelState.IsValid)
                 return View(game);
 
@@ -163,6 +172,11 @@ namespace Client.Controllers
         // GET: GameController/Delete/5
         public async Task<IActionResult> Delete(Guid id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             var client = _httpFactory.CreateClient("Api");
             var game = await client.GetFromJsonAsync<GameViewModel>($"api/Game/{id}");
             if (game == null)
@@ -175,6 +189,11 @@ namespace Client.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(Guid id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
             var client = _httpFactory.CreateClient("Api");
             var response = await client.DeleteAsync($"api/Game/{id}");
